requirements.txt
gradio
pandas
numpy
scikit-learn
sqlalchemy
python-dotenv
plotly